Add test cases. Solve 2016 day 1 part 1

+++ src/y2016/d9/p2.rs
@@ -1,5 +1,5 @@
 #[allow(dead_code)]
-pub fn main(input: &str) -> i32 {
+pub fn main(input: &str) -> i64 {
     println!("p1.rs: {}", input.chars().count());
     0
 }
